MacGui: enable and fix more compiler warnings in the Xcode project.
git-svn-id: svn://svn.handbrake.fr/HandBrake/trunk@6772 b64f7644-9d1e-0410-96f1-a4d463321fa5

+++ macosx/HBAudioTrackPreset.h
@@ -22,7 +22,7 @@
 @property (nonatomic, readwrite) int bitRate;
 
 @property (nonatomic, readwrite) int gain;
-@property (nonatomic, readwrite) float drc;
+@property (nonatomic, readwrite) double drc;
 
 /**
  *  Arrays of possible options for the track properties.
